Consolidated CredentialView, DateLevelView, and Fields View into one.
Having the views segregated made them seem simpler and easier to handle,
but they began to need to know state information about each other.
Rather than passing around pointers and adding methods to manipulate the
views from other views it made sense to consolidate them into a single
view.

The first empty required field now receives the focus on startup.
