+++ app/logbook/ui/src/main/java/org/phoebus/logbook/ui/write/LogEntryDialog.java
@@ -23,6 +23,7 @@ import org.phoebus.logbook.Logbook;
 import org.phoebus.logbook.Tag;
 import org.phoebus.ui.dialog.DialogHelper;
 
+import javafx.event.ActionEvent;
 import javafx.geometry.Insets;
 import javafx.scene.Node;
 import javafx.scene.control.Button;
@@ -36,6 +37,8 @@ import javafx.scene.layout.VBox;
 
 /**
  * Dialog for making an entry into a log book.
+ * * <p> Username, password, title, and logbooks are all required fields. 
+ * The button to submit entry will be disabled if they are not all set.
  * @author Evan Smith
  */
 public class LogEntryDialog extends Dialog<LogEntry>
@@ -48,12 +51,6 @@ public class LogEntryDialog extends Dialog<LogEntry>
     
     /** Dialog Content */
     private final VBox                    content;
-    
-    /** View handles user credential entry for access to log. */
-    private final CredentialEntryView  credentialEntry;
-    
-    /** View handles displaying of date and log entry level selection. */
-    private final DateLevelView        dateAndLevel;
     
     /** View handles the input for creation of the entry. */
     private final FieldsView      logEntryFields;
@@ -73,12 +70,6 @@ public class LogEntryDialog extends Dialog<LogEntry>
         
         content = new VBox();
         
-        // user name and password label and fields.
-        credentialEntry = new CredentialEntryView(model);
-        
-        // date and level labels, fields, and selectors.
-        dateAndLevel = new DateLevelView(model);
-        
         // title and text labels and fields.
         logEntryFields = new FieldsView(model);
                 
@@ -88,11 +79,11 @@ public class LogEntryDialog extends Dialog<LogEntry>
         // Let the Text Area grow to the bottom.
         VBox.setVgrow(logEntryFields,  Priority.ALWAYS);
 
-        VBox.setMargin(credentialEntry, new Insets(10, 0,  0, 0));
+        //VBox.setMargin(credentialEntry, new Insets(10, 0,  0, 0));
         VBox.setMargin(logEntryFields,  new Insets( 0, 0, 10, 0));
         
         content.setSpacing(10);
-        content.getChildren().addAll(credentialEntry, dateAndLevel, logEntryFields, attachmentsView);
+        content.getChildren().addAll(logEntryFields, attachmentsView);
         
         setTitle("Create Log Book Entry");
         
@@ -112,6 +103,13 @@ public class LogEntryDialog extends Dialog<LogEntry>
         // Bind the submit button's disable property to the inverse of the model's ready to submit property.
         submitButton.disableProperty().bind(model.getReadyToSubmitProperty().not());
         submitButton.setTooltip(new Tooltip("Submit Log Entry"));
+        // Prevent enter from causing log entry submission. We want the button to be clicked.
+        // If the button doesn't have focus, it wasn't clicked.
+        submitButton.addEventFilter(ActionEvent.ACTION, eventFilter -> 
+        {
+            if (!submitButton.isFocused())
+                eventFilter.consume();
+        });
         
         setResultConverter(buttonType ->
         {
